+++ airbyte-lib/airbyte_lib/cache.py
@@ -1,19 +1,23 @@
 
 
-from typing import Iterable
-from airbyte_cdk.models import AirbyteRecordMessage
+from typing import Any, Dict, Iterable, List
+from abc import ABC, abstractmethod
+from airbyte_protocol.models import AirbyteRecordMessage
 
-class InMemoryCache:
+class Cache(ABC):
+    @abstractmethod
+    def write(self, messages: Iterable[AirbyteRecordMessage]):
+        pass
+
+
+class InMemoryCache(Cache):
     """The in-memory cache is accepting airbyte messages and stores them in a dictionary for streams (one list of dicts per stream)."""
 
     def __init__(self):
-        self.streams = {}
+        self.streams: Dict[str, List[Dict[str, Any]]] = {}
     
     def write(self, messages: Iterable[AirbyteRecordMessage]):
         for message in messages:
             if message.stream not in self.streams:
                 self.streams[message.stream] = []
             self.streams[message.stream].append(message.data)
-
-def get_in_memory_cache():
-    return InMemoryCache()
