--- airbyte-lib/airbyte_lib/sync.py
@@ -1,5 +1,0 @@
-from .source import Source
-from .cache import InMemoryCache
-
-def sync(connector: Source, store: InMemoryCache):
-    store.write(connector.read())
