+++ airbyte-lib/pyproject.toml
@@ -7,9 +7,9 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.10"
-airbyte-cdk = ">=0.57.1"
 jsonschema = "3.2.0"
 requests = "^2.31.0"
+airbyte-protocol-models = "^1.0.1"
 
 
 [build-system]
